Add shared ag-protocol crate for structured agent protocol
- Move the structured response model, parser, schema, prompt envelopes, repair prompt, and turn prompt helpers into `crates/ag-protocol`.
- Re-export shared protocol APIs from `agentty` compatibility modules and update the agent stack to use the shared crate.
- Add workspace, pre-commit, postsubmit, and crates.io publish wiring for `ag-protocol`.
- Add regression tests in `crates/ag-protocol/src/envelope.rs` so generated protocol text preserves literal template markers.
- Update architecture docs and backend docs to describe the new protocol boundary and Claude SSO login flow.
