+++ .github/workflows/publish-crates-io.yml
@@ -46,6 +46,11 @@ jobs:
           CRATES_IO_TOKEN: ${{ steps.auth.outputs.token }}
         run: cargo publish -p testty --token "$CRATES_IO_TOKEN"
 
+      - name: Publish `ag-protocol` to crates.io
+        env:
+          CRATES_IO_TOKEN: ${{ steps.auth.outputs.token }}
+        run: cargo publish -p ag-protocol --token "$CRATES_IO_TOKEN"
+
       - name: Publish `agentty` to crates.io
         env:
           CRATES_IO_TOKEN: ${{ steps.auth.outputs.token }}
