+++ .pre-commit-config.yaml
@@ -96,6 +96,14 @@ repos:
         always_run: true
         pass_filenames: false
         stages: [manual]
+      - id: test-ag-protocol-src
+        name: cargo nextest ag-protocol src
+        description: Source tests for `crates/ag-protocol/src`.
+        entry: cargo nextest run --locked --profile ci -p ag-protocol --lib
+        language: system
+        always_run: true
+        pass_filenames: false
+        stages: [manual]
       - id: test-agentty-src
         name: cargo nextest agentty src
         description: Source tests for `crates/agentty/src`.
