+++ Cargo.toml
@@ -18,6 +18,7 @@ homepage = "https://github.com/agentty-xyz/agentty"
 [workspace.dependencies]
 ag-forge = { path = "crates/ag-forge", version = "0.11.1" }
 ag-git = { path = "crates/ag-git", version = "0.11.1" }
+ag-protocol = { path = "crates/ag-protocol", version = "0.11.1" }
 testty = { path = "crates/testty", version = "0.11.1" }
 async-trait = "0.1"
 agent-client-protocol = "1.0.0"
