+++ crates/ag-protocol/GEMINI.md
@@ -1,0 +1,1 @@
+AGENTS.md
