+++ crates/ag-protocol/src/template/protocol_instruction_prompt.md
