+++ crates/agentty/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 ag-forge.workspace = true
 ag-git.workspace = true
+ag-protocol.workspace = true
 agent-client-protocol.workspace = true
 async-trait.workspace = true
 arboard.workspace = true
