+++ crates/agentty/src/infra/agent/protocol.rs
@@ -1,18 +1,7 @@
-//! Structured response protocol subsystem router.
-//!
-//! This parent module intentionally stays router-only:
-//! it exposes the child modules and re-exports the public protocol API.
+//! Compatibility re-exports for shared structured response protocol APIs.
 
-mod model;
-mod parse;
-mod schema;
-
-pub use model::ProtocolRequestProfile;
-pub(crate) use model::{AgentResponse, AgentResponseSummary};
-pub(crate) use parse::{
+pub(crate) use ag_protocol::{
+    AgentResponse, AgentResponseSummary, ProtocolRequestProfile, agent_response_output_schema,
+    agent_response_output_schema_json, build_protocol_repair_prompt,
     format_protocol_parse_debug_details, normalize_turn_response, parse_agent_response_strict,
-};
-pub(crate) use schema::{
-    agent_response_json_schema_json, agent_response_output_schema,
-    agent_response_output_schema_json,
 };
